Squash unused comments

+++ Src/lcddraw.c
@@ -11,7 +11,7 @@
 
 int getDigits(uint32_t data)
 {
-	// We used floorf and log10f to take advantage of FPU
+	// We used floorf and log10f to take advantage of STM32F401RE FPU
 	return (int) floorf(log10f(data)) + 1;	// floorf arrotonda per difetto
 											// int tra parentesi è un cast
 											// serve a dire al compilatore di trattare il valore come intero
@@ -23,20 +23,17 @@ int getDigits(uint32_t data)
 
 void DrawFirstRow(const uint64_t secondsSinceStartup, const uint64_t potvalue, const uint64_t counter, const uint8_t unitMeasure)
 {
-	//uint8_t mappedValue = map(potentiometerValue, 0, 4096, 0, 100);
 	uint8_t digits = getDigits(potvalue);
 	uint64_t dataToShow = 0;
 
 	switch(unitMeasure) {
 	case UNIT_CPM:
-		// dataToPrint = 0;
 		StampaStringaSuLCD(ROW_SHORT_COUNTER, N_COLONNE - 4 - digits, "CPM~");
 		dataToShow = counter;
 		break;
 	case UNIT_nSM:
-		// dataToPrint = 0;
 		StampaStringaSuLCD(ROW_SHORT_COUNTER, N_COLONNE - 5 - digits, "nS/m~"); // TODO: fix cast
-		dataToShow = (uint64_t) ((float) counter / 3.54);
+		dataToShow = (uint64_t) ((double) counter / 3.54);
 		break;
 	}
